fix: eliminate all remaining clippy warnings (0 warnings)
- Use .iter_mut().flatten() instead of for + if let Some (manual_flatten)
- Replace manual Iterator::find with .find().map() (manual_find)
- Use vec![..] instead of Vec::new() + push (vec_init_then_push)
- Use struct literals with field defaults instead of Default::default() + assignment (field_reassign_with_default)
- Change &mut Vec<T> to &mut [T] for tick_buffs (ptr_arg)

+++ ServerRust/src/actors/player.rs
@@ -1267,10 +1267,10 @@ impl PlayerActor {
         body.extend_from_slice(&self.state.y.to_le_bytes());            // location_y
         body.push(self.state.direction);                                // direction
         body.push(0u8);                                                 // hair
-        body.extend_from_slice(&(self.state.hp as i32).to_le_bytes());  // hp
-        body.extend_from_slice(&(self.state.mp as i32).to_le_bytes());  // mp
-        body.extend_from_slice(&(self.state.experience as i64).to_le_bytes()); // experience
-        body.extend_from_slice(&(self.state.max_experience as i64).to_le_bytes()); // max_experience
+        body.extend_from_slice(&self.state.hp.to_le_bytes());  // hp
+        body.extend_from_slice(&self.state.mp.to_le_bytes());  // mp
+        body.extend_from_slice(&self.state.experience.to_le_bytes()); // experience
+        body.extend_from_slice(&self.state.max_experience.to_le_bytes()); // max_experience
         body.extend_from_slice(&0u16.to_le_bytes());                    // level_effects
         body.push(0u8);                                                 // has_hero=false
         body.push(0u8);                                                 // hero_behaviour=None
